Reorganize folder structure.
