+++ components/asp/cdl/decoder/Variable.py
